add common ubo

+++ Project/Engine/Source/CrashBasicGeometry.cpp
@@ -43,7 +43,7 @@ namespace Crash
             //  get buffer stride
             int stride = _GetBufferStride(type);
             int index = 0;
-            int offset = 0;
+            size_t offset = 0;
             if (BasicGeometry::HasFlag(type,BasicGeometry::DataType::Vertex))
             {
                 RenderSystem::Instance()->addBufferToVertexArray(pack.vao, pack.vbo, index++, 3, stride, (const void*)(offset));
@@ -250,9 +250,7 @@ namespace Crash
 
     void BasicGeometry::Cube(DataType type, std::vector<float>& vertices)
     {
-        // 立方体每个面独立顶点，逆时针顺序，适合背面剔除和法线贴图
         std::array<_Vertex, 36> cube = {
-            // 后面 (-Z)
             _Vertex{glm::vec3(-1,-1,-1), glm::vec3(0,0,-1), glm::vec2(0,0), glm::vec3(1,0,0)},
             _Vertex{glm::vec3( 1, 1,-1), glm::vec3(0,0,-1), glm::vec2(1,1), glm::vec3(1,0,0)},
             _Vertex{glm::vec3( 1,-1,-1), glm::vec3(0,0,-1), glm::vec2(1,0), glm::vec3(1,0,0)},
@@ -260,7 +258,6 @@ namespace Crash
             _Vertex{glm::vec3(-1, 1,-1), glm::vec3(0,0,-1), glm::vec2(0,1), glm::vec3(1,0,0)},
             _Vertex{glm::vec3( 1, 1,-1), glm::vec3(0,0,-1), glm::vec2(1,1), glm::vec3(1,0,0)},
 
-            // 前面 (+Z)
             _Vertex{glm::vec3(-1,-1, 1), glm::vec3(0,0,1), glm::vec2(0,0), glm::vec3(-1,0,0)},
             _Vertex{glm::vec3( 1,-1, 1), glm::vec3(0,0,1), glm::vec2(1,0), glm::vec3(-1,0,0)},
             _Vertex{glm::vec3( 1, 1, 1), glm::vec3(0,0,1), glm::vec2(1,1), glm::vec3(-1,0,0)},
@@ -268,7 +265,6 @@ namespace Crash
             _Vertex{glm::vec3( 1, 1, 1), glm::vec3(0,0,1), glm::vec2(1,1), glm::vec3(-1,0,0)},
             _Vertex{glm::vec3(-1, 1, 1), glm::vec3(0,0,1), glm::vec2(0,1), glm::vec3(-1,0,0)},
 
-            // 左面 (-X)
             _Vertex{glm::vec3(-1,-1,-1), glm::vec3(-1,0,0), glm::vec2(0,0), glm::vec3(0,0,-1)},
             _Vertex{glm::vec3(-1,-1, 1), glm::vec3(-1,0,0), glm::vec2(1,0), glm::vec3(0,0,-1)},
             _Vertex{glm::vec3(-1, 1, 1), glm::vec3(-1,0,0), glm::vec2(1,1), glm::vec3(0,0,-1)},
@@ -276,7 +272,6 @@ namespace Crash
             _Vertex{glm::vec3(-1, 1, 1), glm::vec3(-1,0,0), glm::vec2(1,1), glm::vec3(0,0,-1)},
             _Vertex{glm::vec3(-1, 1,-1), glm::vec3(-1,0,0), glm::vec2(0,1), glm::vec3(0,0,-1)},
 
-            // 右面 (+X)
             _Vertex{glm::vec3( 1,-1,-1), glm::vec3(1,0,0), glm::vec2(0,0), glm::vec3(0,0,1)},
             _Vertex{glm::vec3( 1, 1, 1), glm::vec3(1,0,0), glm::vec2(1,1), glm::vec3(0,0,1)},
             _Vertex{glm::vec3( 1,-1, 1), glm::vec3(1,0,0), glm::vec2(1,0), glm::vec3(0,0,1)},
@@ -284,7 +279,6 @@ namespace Crash
             _Vertex{glm::vec3( 1, 1,-1), glm::vec3(1,0,0), glm::vec2(0,1), glm::vec3(0,0,1)},
             _Vertex{glm::vec3( 1, 1, 1), glm::vec3(1,0,0), glm::vec2(1,1), glm::vec3(0,0,1)},
 
-            // 下 (-Y)
             _Vertex{glm::vec3(-1,-1,-1), glm::vec3(0,-1,0), glm::vec2(0,1), glm::vec3(1,0,0)},
             _Vertex{glm::vec3( 1,-1,-1), glm::vec3(0,-1,0), glm::vec2(1,1), glm::vec3(1,0,0)},
             _Vertex{glm::vec3( 1,-1, 1), glm::vec3(0,-1,0), glm::vec2(1,0), glm::vec3(1,0,0)},
@@ -292,7 +286,6 @@ namespace Crash
             _Vertex{glm::vec3( 1,-1, 1), glm::vec3(0,-1,0), glm::vec2(1,0), glm::vec3(1,0,0)},
             _Vertex{glm::vec3(-1,-1, 1), glm::vec3(0,-1,0), glm::vec2(0,0), glm::vec3(1,0,0)},
 
-            // 上 (+Y)
             _Vertex{glm::vec3(-1, 1,-1), glm::vec3(0,1,0), glm::vec2(0,1), glm::vec3(1,0,0)},
             _Vertex{glm::vec3( 1, 1, 1), glm::vec3(0,1,0), glm::vec2(1,0), glm::vec3(1,0,0)},
             _Vertex{glm::vec3( 1, 1,-1), glm::vec3(0,1,0), glm::vec2(1,1), glm::vec3(1,0,0)},
@@ -331,34 +324,27 @@ namespace Crash
 
     void BasicGeometry::Cube(DataType type, std::vector<float>& vertices, std::vector<unsigned int>& indices)
     {
-        // 立方体24顶点+36索引，逆时针顺序，适合背面剔除
         std::array<_Vertex, 24> cubeVertices = {
-            // 前面 (+Z)
             _Vertex{glm::vec3(-1,-1, 1), glm::vec3(0,0,1), glm::vec2(0,0), glm::vec3(1,0,0)}, // 0
             _Vertex{glm::vec3( 1,-1, 1), glm::vec3(0,0,1), glm::vec2(1,0), glm::vec3(1,0,0)}, // 1
             _Vertex{glm::vec3( 1, 1, 1), glm::vec3(0,0,1), glm::vec2(1,1), glm::vec3(1,0,0)}, // 2
             _Vertex{glm::vec3(-1, 1, 1), glm::vec3(0,0,1), glm::vec2(0,1), glm::vec3(1,0,0)}, // 3
-            // 后面 (-Z)
             _Vertex{glm::vec3( 1,-1,-1), glm::vec3(0,0,-1), glm::vec2(0,0), glm::vec3(-1,0,0)}, // 4
             _Vertex{glm::vec3(-1,-1,-1), glm::vec3(0,0,-1), glm::vec2(1,0), glm::vec3(-1,0,0)}, // 5
             _Vertex{glm::vec3(-1, 1,-1), glm::vec3(0,0,-1), glm::vec2(1,1), glm::vec3(-1,0,0)}, // 6
             _Vertex{glm::vec3( 1, 1,-1), glm::vec3(0,0,-1), glm::vec2(0,1), glm::vec3(-1,0,0)}, // 7
-            // 左面 (-X)
             _Vertex{glm::vec3(-1,-1,-1), glm::vec3(-1,0,0), glm::vec2(0,0), glm::vec3(0,0,-1)}, // 8
             _Vertex{glm::vec3(-1,-1, 1), glm::vec3(-1,0,0), glm::vec2(1,0), glm::vec3(0,0,-1)}, // 9
             _Vertex{glm::vec3(-1, 1, 1), glm::vec3(-1,0,0), glm::vec2(1,1), glm::vec3(0,0,-1)}, //10
             _Vertex{glm::vec3(-1, 1,-1), glm::vec3(-1,0,0), glm::vec2(0,1), glm::vec3(0,0,-1)}, //11
-            // 右面 (+X)
             _Vertex{glm::vec3( 1,-1, 1), glm::vec3(1,0,0), glm::vec2(0,0), glm::vec3(0,0,1)}, //12
             _Vertex{glm::vec3( 1,-1,-1), glm::vec3(1,0,0), glm::vec2(1,0), glm::vec3(0,0,1)}, //13
             _Vertex{glm::vec3( 1, 1,-1), glm::vec3(1,0,0), glm::vec2(1,1), glm::vec3(0,0,1)}, //14
             _Vertex{glm::vec3( 1, 1, 1), glm::vec3(1,0,0), glm::vec2(0,1), glm::vec3(0,0,1)}, //15
-            // 上面 (+Y)
             _Vertex{glm::vec3(-1, 1, 1), glm::vec3(0,1,0), glm::vec2(0,0), glm::vec3(1,0,0)}, //16
             _Vertex{glm::vec3( 1, 1, 1), glm::vec3(0,1,0), glm::vec2(1,0), glm::vec3(1,0,0)}, //17
             _Vertex{glm::vec3( 1, 1,-1), glm::vec3(0,1,0), glm::vec2(1,1), glm::vec3(1,0,0)}, //18
             _Vertex{glm::vec3(-1, 1,-1), glm::vec3(0,1,0), glm::vec2(0,1), glm::vec3(1,0,0)}, //19
-            // 下面 (-Y)
             _Vertex{glm::vec3(-1,-1,-1), glm::vec3(0,-1,0), glm::vec2(0,0), glm::vec3(1,0,0)}, //20
             _Vertex{glm::vec3( 1,-1,-1), glm::vec3(0,-1,0), glm::vec2(1,0), glm::vec3(1,0,0)}, //21
             _Vertex{glm::vec3( 1,-1, 1), glm::vec3(0,-1,0), glm::vec2(1,1), glm::vec3(1,0,0)}, //22
@@ -393,17 +379,11 @@ namespace Crash
         }
 
         std::array<unsigned int, 36> cubeIndices = {
-            // 前面
             0, 1, 2, 0, 2, 3,
-            // 后面
             4, 5, 6, 4, 6, 7,
-            // 左面
             8, 9,10, 8,10,11,
-            // 右面
            12,13,14,12,14,15,
-            // 上面
            16,17,18,16,18,19,
-            // 下面
            20,21,22,20,22,23
         };
         indices.insert(indices.end(), cubeIndices.begin(), cubeIndices.end());
